% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refs.R
\name{prep_ref_hu}
\alias{prep_ref_hu}
\title{prep human reference tables based on IMGT database}
\usage{
prep_ref_hu(
  url_imgt = "https://raw.githubusercontent.com/ANHIG/IMGTHLA/Latest/fasta/"
)
}
\arguments{
\item{url_imgt, }{string, github link of IMGT database}
}
\value{
data frame, MHC II sequence of officially named alleles
}
\description{
prep human reference tables based on IMGT database
}
\examples{
\donttest{
out <- prep_ref_hu()
}
}
